api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
torch==2.3.0
torchvision==0.18.0
timm==1.0.9
pillow==10.3.0
huggingface_hub==0.24.0
pydantic==2.7.0
python-multipart==0.0.9